adding user entity, tweaking tests to fix
Former-commit-id: 316dce47b53c368dd247584dd75ca5778fa035f4
Former-commit-id: c9263abeccb39ed10e055400baf432de7189d48a

+++ concrete/src/Multilingual/Page/Section/Section.php
@@ -369,12 +369,14 @@ class Section extends Page
 
     public static function unregisterPage($page)
     {
+        $entity = static::getSectionEntity($page->getCollectionID());
         $db = Database::get();
-        $em = $db->getEntityManager();
-        $em->remove(static::getSectionEntity($page->getCollectionID()));
-        $em->flush();
-
-        $db->Execute('delete from MultilingualPageRelations where cID = ?', array($page->getCollectionID()));
+        if (is_object($entity)) {
+            $em = $db->getEntityManager();
+            $em->remove($entity);
+            $em->flush();
+            $db->Execute('delete from MultilingualPageRelations where cID = ?', array($page->getCollectionID()));
+        }
     }
 
     public static function registerMove($page, $oldParent, $newParent)
